move Coordinates, DirectionExtensions, and Position files into Positions folder, same for their unit tests
