+++ MarsRover.Tests/Models/Plateaus/RectangularPlateauTests.cs
@@ -1,5 +1,5 @@
 ﻿using MarsRover.Models.Plateaus;
-using MarsRover.Models.Positions;
+using MarsRover.Models.Positions.Elementals;
 using MarsRover.Models.Vehicles;
 
 namespace MarsRover.Tests.Models.Plateaus;
