+++ MarsRover.Tests/Models/Positions/PositionStringConverterTests.cs
@@ -1,4 +1,5 @@
 ﻿using MarsRover.Models.Positions;
+using MarsRover.Models.Positions.Elementals;
 
 namespace MarsRover.Tests.Models.Positions;
