+++ MarsRover.Tests/Models/Vehicles/RoverTests.cs
@@ -1,6 +1,6 @@
 ﻿using MarsRover.Models.Instructions;
 using MarsRover.Models.Plateaus;
-using MarsRover.Models.Positions;
+using MarsRover.Models.Positions.Elementals;
 using MarsRover.Models.Vehicles;
 
 namespace MarsRover.Tests.Models.Vehicles;
