+++ MarsRover/AppUI/AskUser.cs
@@ -1,6 +1,7 @@
 ﻿using MarsRover.Models.Instructions;
 using MarsRover.Models.Plateaus;
 using MarsRover.Models.Positions;
+using MarsRover.Models.Positions.Elementals;
 using MarsRover.Models.Vehicles;
 
 namespace MarsRover.AppUI;
