+++ MarsRover/Models/Plateaus/Containers/ObstaclesContainer.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.ObjectModel;
-using MarsRover.Models.Positions;
+using MarsRover.Models.Positions.Elementals;
 
 namespace MarsRover.Models.Plateaus.Containers;
