+++ MarsRover/Models/Vehicles/VehicleBase.cs
@@ -1,6 +1,6 @@
 ﻿using MarsRover.Models.Instructions;
 using MarsRover.Models.Plateaus;
-using MarsRover.Models.Positions;
+using MarsRover.Models.Positions.Elementals;
 
 namespace MarsRover.Models.Vehicles;
