+++ MarsRover/Program.cs
@@ -2,6 +2,7 @@
 using MarsRover.Models.Instructions;
 using MarsRover.Models.Plateaus;
 using MarsRover.Models.Positions;
+using MarsRover.Models.Positions.Elementals;
 using MarsRover.Models.Vehicles;
 
 IPositionStringConverter positionStringConverter = new PositionStringConverter();
